feat: Extract media type from asset again
