+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This file contains a chronological list of all major changes made to this crate.
 - Extract asset from proces module
 - Combine the asset and media_type modules
 - Rename process to proc, which sounds less like an OS process
+- Extract media type from asset again
 
 ### 🐛 Fixes
