+++ src/lib.rs
@@ -1,2 +1,4 @@
+mod media_type;
+pub use media_type::{MediaCategory, MediaType};
 pub mod proc;
 pub mod tool;
